requirements.txt
gitpython==3.1.24
progressbar2==3.53.1
